requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.0
torch-spline-conv>=1.2.2
dgl>=1.1.0
dgllife>=0.3.2

# Data processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Molecular processing
rdkit>=2023.3.1
biopython>=1.81
pymol-open-source>=2.5.0

# Protein embeddings
fair-esm>=2.0.0
transformers>=4.30.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# ML tools
optuna>=3.1.0
wandb>=0.15.0
tensorboard>=2.12.0

# File handling
parquet>=1.3.1
pyarrow>=12.0.0
h5py>=3.8.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.3
rich>=13.3.0

# Development
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
jupyter>=1.0.0
ipykernel>=6.22.0